+++ rr/aaaa.js
@@ -38,7 +38,8 @@ export default class AAAA extends RR {
   }
 
   /******  IMPORTERS   *******/
-  fromTinydns (str) {
+  fromTinydns (opts) {
+    const str = opts.tinyline
     let fqdn, ip, n, rdata, ttl, ts, loc
 
     switch (str[0]) {
@@ -67,9 +68,9 @@ export default class AAAA extends RR {
     })
   }
 
-  fromBind (str) {
+  fromBind (opts) {
     // test.example.com  3600  IN  AAAA  ...
-    const [ owner, ttl, c, type, ip ] = str.split(/\s+/)
+    const [ owner, ttl, c, type, ip ] = opts.bindline.split(/\s+/)
     return new AAAA({
       owner,
       ttl    : parseInt(ttl, 10),