+++ rr/dname.js
@@ -39,24 +39,24 @@ export default class DNAME extends RR {
   }
 
   /******  IMPORTERS   *******/
-  fromTinydns (str) {
+  fromTinydns (opts) {
     // DNAME via generic, :fqdn:n:rdata:ttl:timestamp:lo
-    const [ fqdn, n, rdata, ttl, ts, loc ] = str.substring(1).split(':')
+    const [ fqdn, n, rdata, ttl, ts, loc ] = opts.tinyline.substring(1).split(':')
     if (n != 39) throw new Error('DNAME fromTinydns, invalid n')
 
     return new DNAME({
       type     : 'DNAME',
       owner    : this.fullyQualify(fqdn),
-      target   : `${TINYDNS.unpackDomainName(rdata)}.`,
+      target   : (TINYDNS.unpackDomainName(rdata))[0],
       ttl      : parseInt(ttl, 10),
       timestamp: ts,
       location : loc !== '' && loc !== '\n' ? loc : '',
     })
   }
 
-  fromBind (str) {
+  fromBind (opts) {
     // test.example.com  3600  IN  DNAME  ...
-    const [ owner, ttl, c, type, target ] = str.split(/\s+/)
+    const [ owner, ttl, c, type, target ] = opts.bindline.split(/\s+/)
     return new DNAME({
       owner,
       ttl  : parseInt(ttl, 10),