+++ rr/dnskey.js
@@ -1,6 +1,8 @@
 
 import RR from '../rr.js'
 
+import * as TINYDNS from '../lib/tinydns.js'
+
 export default class DNSKEY extends RR {
   constructor (opts) {
     super(opts)
@@ -56,10 +58,10 @@ export default class DNSKEY extends RR {
 
   /******  IMPORTERS   *******/
 
-  fromBind (str) {
+  fromBind (opts) {
     // test.example.com  3600  IN  DNSKEY Flags Protocol Algorithm PublicKey
-    const match = str.match(/^([^\s]+)\s+([0-9]+)\s+(\w+)\s+(\w+)\s+([0-9]+)\s+([0-9]+)\s+([0-9]+)\s+\s*(.*?)\s*$/)
-    if (!match) throw new Error(`unable to parse DNSKEY: ${str}`)
+    const match = opts.bindline.match(/^([^\s]+)\s+([0-9]+)\s+(\w+)\s+(\w+)\s+([0-9]+)\s+([0-9]+)\s+([0-9]+)\s+\s*(.*?)\s*$/)
+    if (!match) throw new Error(`unable to parse DNSKEY: ${opts.bindline}`)
     const [ owner, ttl, c, type, flags, protocol, algorithm, publickey ] = match.slice(1)
 
     return new DNSKEY({
@@ -74,5 +76,35 @@ export default class DNSKEY extends RR {
     })
   }
 
+  fromTinydns (opts) {
+    const [ fqdn, n, rdata, ttl, ts, loc ] = opts.tinyline.substring(1).split(':')
+    if (n != 48) throw new Error('DNSKEY fromTinydns, invalid n')
+
+    const bytes = Buffer.from(TINYDNS.octalToChar(rdata), 'binary')
+
+    return new DNSKEY({
+      owner      : this.fullyQualify(fqdn),
+      ttl        : parseInt(ttl, 10),
+      type       : 'DNSKEY',
+      flags      : bytes.readUInt16BE(0),
+      protocol   : bytes.readUInt8(2),
+      'algorithm': bytes.readUInt8(3),
+      'publickey': bytes.slice(4).toString(),
+      timestamp  : ts,
+      location   : loc !== '' && loc !== '\n' ? loc : '',
+    })
+  }
+
   /******  EXPORTERS   *******/
+
+  toTinydns () {
+    const dataRe = new RegExp(/[\r\n\t:\\/]/, 'g')
+
+    return this.getTinydnsGeneric(
+      TINYDNS.UInt16toOctal(this.get('flags')) +
+      TINYDNS.UInt8toOctal(this.get('protocol')) +
+      TINYDNS.UInt8toOctal(this.get('algorithm')) +
+      TINYDNS.escapeOctal(dataRe, this.get('publickey'))
+    )
+  }
 }