+++ test/dnskey.js
@@ -14,7 +14,7 @@ const validRecords = [
     algorithm: 5,
     publickey: `( AQPSKmynfzW4kyBv015MUG2DeIQ3 Cbl+BBZH4b/0PY1kxkmvHjcZc8no kfzj31GajIQKY+5CptLr3buXA10h WqTkF7H6RfoRqXQeogmMHfpftf6z Mv1LyBUgia7za6ZEzOJBOztyvhjL 742iU/TpPSEDhm2SNKLijfUppn1U aNvv4w== )`,
     testB    : 'example.com.\t3600\tIN\tDNSKEY\t256\t3\t5\t( AQPSKmynfzW4kyBv015MUG2DeIQ3 Cbl+BBZH4b/0PY1kxkmvHjcZc8no kfzj31GajIQKY+5CptLr3buXA10h WqTkF7H6RfoRqXQeogmMHfpftf6z Mv1LyBUgia7za6ZEzOJBOztyvhjL 742iU/TpPSEDhm2SNKLijfUppn1U aNvv4w== )\n',
-    testT    : ':dnskey.example.com:33:\\000\\001\\000\\000\\003\\341\\004mail\\007example\\003com\\000:3600::\n',
+    testT    : ':example.com:48:\\001\\000\\003\\005( AQPSKmynfzW4kyBv015MUG2DeIQ3 Cbl+BBZH4b\\0570PY1kxkmvHjcZc8no kfzj31GajIQKY+5CptLr3buXA10h WqTkF7H6RfoRqXQeogmMHfpftf6z Mv1LyBUgia7za6ZEzOJBOztyvhjL 742iU\\057TpPSEDhm2SNKLijfUppn1U aNvv4w== ):3600::\n',
   },
 ]
 
@@ -37,8 +37,8 @@ describe('DNSKEY record', function () {
   base.getTypeId(DNSKEY, 48)
 
   base.toBind(DNSKEY, validRecords)
-  // base.toTinydns(DNSKEY, validRecords)
+  base.toTinydns(DNSKEY, validRecords)
 
   base.fromBind(DNSKEY, validRecords)
-  // base.fromTinydns(DNSKEY, validRecords)
+  base.fromTinydns(DNSKEY, validRecords)
 })