+++ test/naptr.js
@@ -16,7 +16,7 @@ const validRecords = [
     regexp     : '!^urn:cid:.+@([^\\.]+\\.)(.*)$!\x02!i',
     replacement: '.',
     testB      : 'cid.urn.arpa.\t86400\tIN\tNAPTR\t100\t10\t""\t""\t"!^urn:cid:.+@([^\\.]+\\.)(.*)$!\x02!i"\t.\n',
-    testT      : ':cid.urn.arpa:35:\\000\\144\\000\\012\\000\\000\\040!^urn\\072cid\\072.+@([^\\134.]+\\134.)(.*)$!\x02!i\\001.``000:86400::\n',
+    testT      : ':cid.urn.arpa:35:\\000\\144\\000\\012\\000\\000\\040!^urn\\072cid\\072.+@([^\\134.]+\\134.)(.*)$!\x02!i\\001.\\000:86400::\n',
   },
 ]
 
@@ -36,5 +36,5 @@ describe('NAPTR record', function () {
   base.toTinydns(NAPTR, validRecords)
 
   base.fromBind(NAPTR, validRecords)
-  // base.fromTinydns(NAPTR, validRecords)
+  base.fromTinydns(NAPTR, validRecords)
 })