+++ test/tinydns.js
@@ -5,6 +5,27 @@ import * as TINYDNS from '../lib/tinydns.js'
 
 describe('TINYDNS', function () {
 
+  const b64cases = {
+    '0sAQPeOwAGDPLrDebL1q5Lg8XW9B/d9MnxqlzIYKXhvZPWEHNYGP7AwART/tmkeDNn7HPMtgM6GIwQ4p0KGLfSRoUKbjtPlRVeWYLbsnNXeFU5bchyYef0efYiKlxZdo':
+    '\\322\\300\\020\\075\\343\\260\\000\\140\\317.\\260\\336l\\275j\\344\\270\\074\\135oA\\375\\337L\\237\\032\\245\\314\\206\\012\\136\\033\\331\\075a\\0075\\201\\217\\354\\014\\000E\\077\\355\\232G\\2036\\176\\307\\074\\313\\1403\\241\\210\\301\\016\\051\\320\\241\\213\\175\\044hP\\246\\343\\264\\371QU\\345\\230-\\273\\0475w\\205S\\226\\334\\207\\046\\036\\177G\\237b\\042\\245\\305\\227h',
+  }
+
+  describe('base64toOctal', function () {
+    for (const c in b64cases) {
+      it('octal escapes a base64 encoded string', async function () {
+        assert.strictEqual(TINYDNS.base64toOctal(c), b64cases[c])
+      })
+    }
+  })
+
+  describe('octalToBase64', function () {
+    for (const c in b64cases) {
+      it('converts octal escaped to base64 encoded string', async function () {
+        assert.deepStrictEqual(TINYDNS.octalToBase64(b64cases[c]), c)
+      })
+    }
+  })
+
   describe('escapeOctal', function () {
 
     const rdataRe = new RegExp(/[\r\n\t:\\/]/, 'g')
@@ -21,10 +42,14 @@ describe('TINYDNS', function () {
 
     for (const c of Object.keys(specialChars)) {
       it(`escapes tinydns rdata special char ${c}`, function () {
-        const e = TINYDNS.escapeOctal(rdataRe, c)
-        assert.strictEqual(e, specialChars[c])
+        assert.strictEqual(TINYDNS.escapeOctal(rdataRe, c), specialChars[c])
       })
     }
+  })
+
+  describe('octaltoChar', function () {
+    it('converts a string of octal escapes to characters', async () => {
+    })
   })
 
   describe('octalToHex', function () {
@@ -70,5 +95,35 @@ describe('TINYDNS', function () {
     it('converts escaped octal to 32-bit integer', function () {
       assert.strictEqual(TINYDNS.octalToUInt32('\\145\\276\\224\\150'), 1706988648)
     })
+  })
+
+  describe('unpackDomainName', function () {
+    it(`extracts domain name from wire format`, async () => {
+      const r = TINYDNS.unpackDomainName('\\006sipdir\\006online\\004lync\\003com\\000')
+      assert.strictEqual(r[0], 'sipdir.online.lync.com.')
+      assert.strictEqual(r[1], 40)
+    })
+  })
+
+  const cases = {
+    '0.0.0.0'        : '\\000\\000\\000\\000',
+    '192.0.2.127'    : '\\300\\000\\002\\177',
+    '255.255.255.255': '\\377\\377\\377\\377',
+  }
+
+  describe('ipv4toOctal', function () {
+    for (const c in cases) {
+      it(`converts dotted quad IPv4 to octal escaped integer ${c}`, async () => {
+        assert.strictEqual(TINYDNS.ipv4toOctal(c), cases[c])
+      })
+    }
+  })
+
+  describe('octalToIPv4', function () {
+    for (const c in cases) {
+      it(`converts octal escaped integer to dotted quad IPv4 ${cases[c]}`, async () => {
+        assert.strictEqual(TINYDNS.octalToIPv4(cases[c]), c)
+      })
+    }
   })
 })